refactoring directory structure
회원가입 로직 구현
