+++ src/main/java/com/mda/imirror/domain/request/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.mda.imirror.domain;
+package com.mda.imirror.domain.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -11,25 +11,19 @@ import java.util.List;
 @Setter
 public class MemberRegisterRequest {
     @NotBlank
+    private String memberId;
+    @NotBlank
+    private String memberPassword;
+    @NotBlank
     @Length(max = 100)
     private String memberName;
     @NotBlank
-    @Length(max = 1)
     private Boolean memberGender;
     @NotBlank
     @Length(max = 3)
     private Integer memberAge;
-    @NotBlank
     private Double memberHeight;
-    @NotBlank
     private Double memberWeight;
     @NotBlank
-    @Length(max = 1)
     private Boolean personalInfoConsent;
-    @NotBlank
-    @Length(max = 10)
-    private String role;
-
-    private List<Rom> memberPkRoms;
-    private List<Balance> memberPkBalances;
 }
