+++ src/main/java/com/mda/imirror/domain/response/MemberRegisterResponse.java
@@ -1,0 +1,16 @@
+package com.mda.imirror.domain.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MemberRegisterResponse {
+    private String memberPk;
+    private String memberName;
+
+}
